refactor(Block) : Move Block related component in common folder
