+++ src/components/common/Block/BlockPopUp.tsx
@@ -1,9 +1,9 @@
 import React from "react";
 import { getBlockType } from "../../../typedef/common/common.types";
 import "./css/BlockPopUp.css";
-import CommentScrollViewContainer from "../containers/CommentScrollViewContainer";
+import CommentScrollViewContainer from "./containers/CommentScrollViewContainer";
 import images from "../../../assets/images";
-import PopUpVoteViewContainer from "../containers/PopUpVoteViewContainer";
+import PopUpVoteViewContainer from "./containers/PopUpVoteViewContainer";
 
 type Props = {
   blockDetail: getBlockType;
