+++ src/components/common/Block/VoteView.tsx
