+++ src/components/common/Block/containers/BlockPopUpContainer.tsx
@@ -1,17 +1,17 @@
 import React, { useCallback, useEffect, useRef, useState } from "react";
-import BlockPopUp from "../components/BlockPopUp";
+import BlockPopUp from "../BlockPopUp";
 import {
   BasicAPIResponseType,
   getBlockType,
   LoginTokenType,
-} from "../../../typedef/common/common.types";
+} from "../../../../typedef/common/common.types";
 import {
   apiOrigin,
   apiRoute,
   requestDelete,
   requestPost,
-} from "../../../lib/api/api";
-import useAuth from "../../../hooks/Auth/useAuth";
+} from "../../../../lib/api/api";
+import useAuth from "../../../../hooks/Auth/useAuth";
 
 type Props = {
   blockDetail: getBlockType;
