+++ src/components/common/Block/containers/PopUpVoteViewContainer.tsx
@@ -1,11 +1,11 @@
 import React, { useEffect, useState } from "react";
-import useAuth from "../../../hooks/Auth/useAuth";
-import { apiOrigin, apiRoute, requestPost } from "../../../lib/api/api";
+import useAuth from "../../../../hooks/Auth/useAuth";
+import { apiOrigin, apiRoute, requestPost } from "../../../../lib/api/api";
 import {
   BasicAPIResponseType,
   getBlockType,
-} from "../../../typedef/common/common.types";
-import VoteView from "../components/VoteView";
+} from "../../../../typedef/common/common.types";
+import VoteView from "../VoteView";
 
 type Props = {
   votedIndex: number;
