+++ src/components/common/Block/css/BlockPopUp.scss
