+++ src/components/common/Block/css/VoteView.css
