+++ src/components/common/Block/css/VoteView.scss
