+++ src/components/common/Block/css/imagePopUp.css
@@ -16,7 +16,7 @@
   width: 16px;
   height: 16px;
   -o-object-fit: contain;
-     object-fit: contain;
+  object-fit: contain;
 }
 
 #root .image-popup-wrap .image-area {
@@ -29,6 +29,6 @@
   width: 100%;
   height: 100%;
   -o-object-fit: contain;
-     object-fit: contain;
+  object-fit: contain;
 }
 /*# sourceMappingURL=imagePopUp.css.map */
